move common loading code out of csv_helper/loader
